Minimal set of fixes making project building on java 11+

+++ smartparam-engine/src/test/java/org/smartparam/engine/report/query/ParamQueryIntegrationTest.java
@@ -159,7 +159,7 @@ public class ParamQueryIntegrationTest {
 
         // then
         assertThat(paramValue).hasSize(1);
-        assertThat(paramValue.get()).isEqualTo(50L);
+        assertThat((long) paramValue.get()).isEqualTo(50L);
     }
 
     private static class HigherValueChooser extends DecodedValueChooser {
